# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SVAES API",
    "description": "Automatic Software Delivery Verification System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List Organizations",
        "description": "Endpoint para listar las organizaciones. Solo los usuarios con rol ADMIN pueden acceder a esta informaci\u00f3n.\n\nAtributos:\n    - skip: int - N\u00famero de registros a omitir para paginaci\u00f3n.\n    - limit: int - N\u00famero m\u00e1ximo de registros a retornar.\n    - current_user: Usuario autenticado con permisos del token JWT.\n    - service: IOrganizationService - El servicio de organizaciones, inyectado mediante dependencias.\n\nRetorna:\n    - Una lista de diccionarios con la informaci\u00f3n de cada organizaci\u00f3n.\n    - Lanza HTTPException con status 403 si el usuario no tiene rol ADMIN.\n    - Lanza HTTPException con status 500 para cualquier error inesperado.",
        "operationId": "list_organizations_api_v1_organizations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create Organization",
        "description": "Endpoint para crear una nueva organizaci\u00f3n. Solo los usuarios con rol ADMIN pueden crear organizaciones.\n\nAtributos:\n    - payload: OrganizationCreateRequest - El cuerpo de la solicitud con los datos de la organizaci\u00f3n a crear.\n    - current_user: Usuario autenticado con permisos del token JWT.\n    - service: IOrganizationService - El servicio de organizaciones, inyectado mediante dependencias.\n\nRetorna:\n    - Un diccionario con la informaci\u00f3n de la organizaci\u00f3n creada (id, name, slug).\n    - Lanza HTTPException con status 403 si el usuario no tiene rol ADMIN.\n    - Lanza HTTPException con status 409 si hay un error de validaci\u00f3n (e.g., slug ya existe).\n    - Lanza HTTPException con status 500 para cualquier error inesperado.",
        "operationId": "create_organization_api_v1_organizations_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{org_id}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get Organization",
        "description": "Endpoint para obtener los detalles de una organizaci\u00f3n.\n\nAtributos:\n    - org_id: UUID - El ID de la organizaci\u00f3n.\n    - current_user: Usuario autenticado con permisos del token JWT.\n    - service: IOrganizationService - El servicio de organizaciones, inyectado mediante dependencias.\n\nRetorna:\n    - Un diccionario con la informaci\u00f3n de la organizaci\u00f3n (id, name, slug, owner_id, etc.).\n    - Lanza HTTPException con status 403 si el usuario no tiene acceso a la organizaci\u00f3n.\n    - Lanza HTTPException con status 404 si la organizaci\u00f3n no existe.\n    - Lanza HTTPException con status 500 para cualquier error inesperado.",
        "operationId": "get_organization_api_v1_organizations__org_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get Project By Id",
        "operationId": "get_project_by_id_api_v1_projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List Accessible Projects",
        "operationId": "list_accessible_projects_api_v1_projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Size"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{org_id}/projects": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "List Org Projects",
        "operationId": "list_org_projects_api_v1_organizations__org_id__projects_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organizations"
        ],
        "summary": "Create Project",
        "description": "Endpoint para crear un nuevo proyecto dentro de una organizaci\u00f3n.\n\nAtributos:\n    - org_id: UUID - El ID de la organizaci\u00f3n.\n    - payload: ProjectCreateRequest - El cuerpo de la solicitud con los datos del proyecto a crear.\n    - current_user: Usuario autenticado con permisos del token JWT.\n    - service: IOrganizationService - El servicio de organizaciones, inyectado mediante dependencias.\n\nRetorna:\n    - Un diccionario con la informaci\u00f3n del proyecto creado (id, name).\n    - Lanza HTTPException con status 403 si el usuario no tiene acceso a la organizaci\u00f3n.\n    - Lanza HTTPException con status 404 si la organizaci\u00f3n no existe.\n    - Lanza HTTPException con status 409 si hay un error de validaci\u00f3n.\n    - Lanza HTTPException con status 500 para cualquier error inesperado.",
        "operationId": "create_project_api_v1_organizations__org_id__projects_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{org_id}/projects/{project_id}": {
      "get": {
        "tags": [
          "Organizations"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_v1_organizations__org_id__projects__project_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "pattern": "^[a-z0-9-]+$",
            "title": "Slug"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "slug"
        ],
        "title": "OrganizationCreateRequest"
      },
...
      "ProjectCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 500,
            "title": "Description",
            "default": ""
          },
          "profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Profile Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "profile_id"
        ],
        "title": "ProjectCreateRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}